+++ back/src/main/java/com/ll/goohaeyou/notification/application/NotificationEventListener.java
@@ -1,6 +1,5 @@
-package com.ll.goohaeyou.notification.eventListener;
+package com.ll.goohaeyou.notification.application;
 
-import com.ll.goohaeyou.notification.application.NotificationService;
 import com.ll.goohaeyou.global.event.notification.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
